refactor(dts): split jsdoc helpers below line cap
